pyproject.toml
[project]
name = "gpu-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"

# Base dependencies only — minimal runtime
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "celery>=5.6.2",
    "eventlet>=0.40.4",
    "gradio>=6.3.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "setuptools>=80.9.0",
    "transformers>=4.57.3",
]

# Optional dependency groups
[project.optional-dependencies]
api = [
    "flower>=2.0.1",
    "fastapi>=0.128.0",
    "celery>=5.6.2",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.9",
    "kombu>=5.4.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.40.0",
    "httpx>=0.27.0",
]
deepseek-ocr = [
    "addict>=2.4.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "matplotlib>=3.10.8",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ninja>=1.13.0",
    "kernels>=0.11.7",
    
]

gpt-oss = [
        "torch>=2.9.1",
    "transformers>=4.57.3",
        "accelerate>=1.12.0",

]

all = [
    "gpu-server[api,deepseek-ocr,gpt-oss]",
]


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
